Move async_task from util to task
