--- app/src/main/java/com/litechmeg/sabocale/util/KamokuAsyncTask.java
@@ -1,7 +1,0 @@
-package com.litechmeg.sabocale.util;
-
-/**
- * Created by megukanipan on 2015/05/09.
- */
-public class KamokuAsyncTask {
-}
