+++ app/src/main/java/com/litechmeg/sabocale/view/activity/MainActivity.java
@@ -3,7 +3,6 @@ package com.litechmeg.sabocale.view.activity;
 import android.app.AlertDialog;
 import android.app.ProgressDialog;
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.content.res.Configuration;
 import android.graphics.Color;
 import android.os.Bundle;
@@ -32,7 +31,7 @@ import com.litechmeg.sabocale.R;
 import com.litechmeg.sabocale.model.Kamoku;
 import com.litechmeg.sabocale.model.Subject;
 import com.litechmeg.sabocale.model.Term;
-import com.litechmeg.sabocale.util.AttendanceAsyncTask;
+import com.litechmeg.sabocale.task.AttendanceAsyncTask;
 import com.litechmeg.sabocale.util.PrefUtils;
 import com.litechmeg.sabocale.view.adapter.MainTabPagerAdapter;
 import com.litechmeg.sabocale.view.adapter.TermListArrayAdapter;
